[FIX] account: limit attachments discarding to emails only
When attaching an image in the account module, the `_check_and_decode_attachment`
method led to the deletion of non-decodable attachments, causing subsequent access
errors in `IrAttachment._post_add_create` since it was trying to use deleted record

Reproduce
---
- -i account
- Create new draft invoice (without lines) -> Save
- Attach some pdf with invoice
- Try attaching some image to the chatter -> BUG: stuck on uploading

While adapting the tests to test correctly the user endpoints instead of
the internal functions, we realized that some cases are not possible:
* plain text and CSV files will be discarded because we can't memorize
  their mimetype when getting the mail, and we can't guess it either.
* when we are uploading multiple attachments on the journal, we can't
  regroup multiple attachments on the same invoices like we do on the
  mail alias. This could be changed in the future, but the test has only
  been adapted to reflect the current status.
* the xml preface is not mandatory but we were using it to guess the
  mimetype of xml files. We use `guess_mimetype` instead.

opw-4144868

X-original-commit: a2b7fd05c46ca8fe6103d4276ca08b7aaab4fde4

+++ addons/account/models/account_move.py
@@ -3572,6 +3572,11 @@ class AccountMove(models.Model):
                     attachments_by_invoice[attachment] |= invoice
                 else:
                     attachments_by_invoice[attachment] = invoice
+                if not attachment.res_id:
+                    attachment.write({
+                        'res_id': invoice.id,
+                        'res_model': invoice._name,
+                    })
 
         file_data_list = attachments._unwrap_edi_attachments()
         attachments_by_invoice = {}
@@ -5271,8 +5276,9 @@ class AccountMove(models.Model):
         attachments_in_invoices = self.env['ir.attachment']
         for attachment in move_per_decodable_attachment:
             attachments_in_invoices += attachment
-        # Unlink the unused attachments
-        (attachments - attachments_in_invoices).unlink()
+        # Unlink the unused attachments (prevents storing marketing images sent with emails)
+        if self._context.get('from_alias'):
+            (attachments - attachments_in_invoices).unlink()
         return move_per_decodable_attachment
 
     def _creation_subtype(self):
